fix(app-deployment) :: resource 요청 memory/disk를 MB 단위로 변환

+++ src/main/java/madp/appdeployment/domain/infrastructure/client/response/AppRevisionResponseDto.java
@@ -10,7 +10,7 @@ public record AppRevisionResponseDto(
         Integer maxCpu,
 
         @JsonProperty("max_memory")
-        Double maxMemory,
+        Integer maxMemory,
 
         @JsonProperty("max_disk")
         Integer maxDisk,
